deconv reclustering; work in progress
